% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mate.R
\name{mate}
\alias{mate}
\title{Title}
\usage{
mate(
  parent1,
  parent2,
  fitness1 = NULL,
  fitness2 = NULL,
  p_mutation = 0.05,
  mutated_genes = NULL
)
}
\arguments{
\item{parent1}{Vector of genes from parent 1}

\item{parent2}{Vector of genes from parent 2}

\item{fitness1}{Fitness of parent 1. NULL gives equal fitness of both parents, otherwise scale so that a higher value of fitness corresponds to better fitness}

\item{fitness2}{Fitness of parent 2. NULL gives equal fitness of both parents, otherwise scale so that a higher value of fitness corresponds to better fitness}

\item{p_mutation}{Probability of mutation in the genes}

\item{mutated_genes}{New genes to be inserted for mutations}
}
\value{
A vector of a child with genes from parent 1 and 2
}
\description{
Title
}
