% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rhnorm}
\alias{rhnorm}
\title{Random draws from half normal distribution}
\usage{
rhnorm(n, sd = 1)
}
\arguments{
\item{n}{Number of observations}

\item{sd}{standard deviation}
}
\value{
Vector of random draws from half normal distribution
}
\description{
Random draws from half normal distribution
}
\examples{
rhnorm(10)
}
