% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_genetic.R
\name{run_genetic_knn}
\alias{run_genetic_knn}
\title{Run genetic knn}
\usage{
run_genetic_knn(
  train,
  test = NULL,
  formula,
  k,
  gene_dist,
  metric = "MSE",
  metric_type = c("lower_is_better", "higher_is_better"),
  n_generations = 10,
  n_individuals = 100,
  n_discard = 10,
  n_keep = 10,
  standardize = T,
  normalize = F,
  normalize_scores = T,
  timing = T,
  plot_results = T,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Run genetic knn
}
\examples{
run_genetic_knn(x,y,z)
}
